Move math, malloc, and mem out
