+++ c-scape/src/lib.rs
@@ -40,8 +40,6 @@ use alloc::boxed::Box;
 use alloc::format;
 use core::convert::TryInto;
 use core::ffi::c_void;
-#[cfg(not(target_arch = "riscv64"))]
-use core::mem::align_of;
 #[cfg(not(target_os = "wasi"))]
 use core::mem::{size_of, zeroed};
 use core::ptr::{self, null, null_mut};
@@ -57,6 +55,19 @@ mod fs;
 
 // io
 mod io;
+
+// malloc
+mod malloc;
+
+pub use malloc::*;
+
+// math
+mod math;
+
+// mem
+mod mem;
+
+pub use mem::*;
 
 // mmap
 #[cfg(not(target_os = "wasi"))]
@@ -98,220 +109,6 @@ unsafe extern "C" fn __xpg_strerror_r(errnum: c_int, buf: *mut c_char, buflen: u
     out.copy_from_slice(message.as_bytes());
     out[out.len() - 1] = b'\0';
     0
-}
-
-// malloc
-
-#[repr(C)]
-#[derive(Clone, Copy, Debug)]
-struct Tag {
-    size: usize,
-    align: usize,
-}
-
-/// allocates for a given layout, with a tag prepended to the allocation,
-/// to keep track of said layout.
-/// returns null if the allocation failed
-fn tagged_alloc(type_layout: alloc::alloc::Layout) -> *mut u8 {
-    if type_layout.size() == 0 {
-        return core::ptr::null_mut();
-    }
-
-    let tag_layout = alloc::alloc::Layout::new::<Tag>();
-    let tag = Tag {
-        size: type_layout.size(),
-        align: type_layout.align(),
-    };
-    if let Ok((total_layout, offset)) = tag_layout.extend(type_layout) {
-        let total_ptr = unsafe { alloc::alloc::alloc(total_layout) };
-        if total_ptr.is_null() {
-            return total_ptr;
-        }
-
-        let tag_offset = offset - tag_layout.size();
-        unsafe {
-            total_ptr.wrapping_add(tag_offset).cast::<Tag>().write(tag);
-            total_ptr.wrapping_add(offset).cast()
-        }
-    } else {
-        core::ptr::null_mut()
-    }
-}
-
-/// get the layout out of a tagged allocation
-///
-/// #Safety
-/// the given pointer must be a non-null pointer,
-/// gotten from calling `tagged_alloc`
-unsafe fn get_layout(ptr: *mut u8) -> alloc::alloc::Layout {
-    let tag = ptr
-        .wrapping_sub(core::mem::size_of::<Tag>())
-        .cast::<Tag>()
-        .read();
-    alloc::alloc::Layout::from_size_align_unchecked(tag.size, tag.align)
-}
-
-/// get the layout out of a tagged allocation
-///
-/// #Safety
-/// the given pointer must be a non-null pointer,
-/// gotten from calling `tagged_alloc`
-unsafe fn tagged_dealloc(ptr: *mut u8) {
-    let tag_layout = alloc::alloc::Layout::new::<Tag>();
-    let type_layout = get_layout(ptr);
-    if let Ok((total_layout, offset)) = tag_layout.extend(type_layout) {
-        let total_ptr = ptr.wrapping_sub(offset);
-        alloc::alloc::dealloc(total_ptr, total_layout);
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn malloc(size: usize) -> *mut c_void {
-    libc!(libc::malloc(size));
-
-    // TODO: Add `max_align_t` for riscv64 to upstream libc.
-    #[cfg(target_arch = "riscv64")]
-    let layout = alloc::alloc::Layout::from_size_align(size, 16).unwrap();
-    #[cfg(not(target_arch = "riscv64"))]
-    let layout =
-        alloc::alloc::Layout::from_size_align(size, align_of::<libc::max_align_t>()).unwrap();
-    tagged_alloc(layout).cast()
-}
-
-#[no_mangle]
-unsafe extern "C" fn realloc(old: *mut c_void, size: usize) -> *mut c_void {
-    libc!(libc::realloc(old, size));
-
-    if old.is_null() {
-        malloc(size)
-    } else {
-        let old_layout = get_layout(old.cast());
-        if old_layout.size() >= size {
-            return old;
-        }
-
-        let new = malloc(size);
-        memcpy(new, old, core::cmp::min(size, old_layout.size()));
-        tagged_dealloc(old.cast());
-        new
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn calloc(nmemb: usize, size: usize) -> *mut c_void {
-    libc!(libc::calloc(nmemb, size));
-
-    let product = match nmemb.checked_mul(size) {
-        Some(product) => product,
-        None => {
-            set_errno(Errno(libc::ENOMEM));
-            return null_mut();
-        }
-    };
-
-    let ptr = malloc(product);
-    memset(ptr, 0, product)
-}
-
-#[no_mangle]
-unsafe extern "C" fn posix_memalign(
-    memptr: *mut *mut c_void,
-    alignment: usize,
-    size: usize,
-) -> c_int {
-    libc!(libc::posix_memalign(memptr, alignment, size));
-    if !(alignment.is_power_of_two() && alignment % core::mem::size_of::<*const c_void>() == 0) {
-        return rustix::io::Error::INVAL.raw_os_error();
-    }
-
-    let layout = alloc::alloc::Layout::from_size_align(size, alignment).unwrap();
-    let ptr = tagged_alloc(layout).cast();
-
-    *memptr = ptr;
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn free(ptr: *mut c_void) {
-    libc!(libc::free(ptr));
-
-    if ptr.is_null() {
-        return;
-    }
-
-    tagged_dealloc(ptr.cast());
-}
-
-// mem
-
-#[no_mangle]
-unsafe extern "C" fn memcmp(a: *const c_void, b: *const c_void, len: usize) -> c_int {
-    libc!(libc::memcmp(a, b, len));
-
-    compiler_builtins::mem::memcmp(a.cast(), b.cast(), len)
-}
-
-#[no_mangle]
-unsafe extern "C" fn bcmp(a: *const c_void, b: *const c_void, len: usize) -> c_int {
-    // `bcmp` is just an alias for `memcmp`.
-    libc!(libc::memcmp(a, b, len));
-
-    compiler_builtins::mem::bcmp(a.cast(), b.cast(), len)
-}
-
-#[no_mangle]
-unsafe extern "C" fn memcpy(dst: *mut c_void, src: *const c_void, len: usize) -> *mut c_void {
-    libc!(libc::memcpy(dst, src, len));
-
-    compiler_builtins::mem::memcpy(dst.cast(), src.cast(), len).cast()
-}
-
-#[no_mangle]
-unsafe extern "C" fn memmove(dst: *mut c_void, src: *const c_void, len: usize) -> *mut c_void {
-    libc!(libc::memmove(dst, src, len));
-
-    compiler_builtins::mem::memmove(dst.cast(), src.cast(), len).cast()
-}
-
-#[no_mangle]
-unsafe extern "C" fn memset(dst: *mut c_void, fill: c_int, len: usize) -> *mut c_void {
-    libc!(libc::memset(dst, fill, len));
-
-    compiler_builtins::mem::memset(dst.cast(), fill, len).cast()
-}
-
-#[no_mangle]
-unsafe extern "C" fn memchr(s: *const c_void, c: c_int, len: usize) -> *mut c_void {
-    libc!(libc::memchr(s, c, len));
-
-    let slice: &[u8] = slice::from_raw_parts(s.cast(), len);
-    match memchr::memchr(c as u8, slice) {
-        None => null_mut(),
-        Some(i) => s.cast::<u8>().add(i) as *mut c_void,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn memrchr(s: *const c_void, c: c_int, len: usize) -> *mut c_void {
-    libc!(libc::memrchr(s, c, len));
-
-    let slice: &[u8] = slice::from_raw_parts(s.cast(), len);
-    match memchr::memrchr(c as u8, slice) {
-        None => null_mut(),
-        Some(i) => s.cast::<u8>().add(i) as *mut c_void,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn strlen(mut s: *const c_char) -> usize {
-    libc!(libc::strlen(s));
-
-    let mut len = 0;
-    while *s != b'\0' as _ {
-        len += 1;
-        s = s.add(1);
-    }
-    len
 }
 
 // rand
@@ -941,578 +738,6 @@ unsafe extern "C" fn siglongjmp() {
 unsafe extern "C" fn __sigsetjmp() {
     //libc!(libc::__sigsetjmp());
     unimplemented!("__sigsetjmp")
-}
-
-// math
-
-#[no_mangle]
-unsafe extern "C" fn acos(x: f64) -> f64 {
-    libm::acos(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn acosf(x: f32) -> f32 {
-    libm::acosf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn acosh(x: f64) -> f64 {
-    libm::acosh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn acoshf(x: f32) -> f32 {
-    libm::acoshf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn asin(x: f64) -> f64 {
-    libm::asin(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn asinf(x: f32) -> f32 {
-    libm::asinf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn asinh(x: f64) -> f64 {
-    libm::asinh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn asinhf(x: f32) -> f32 {
-    libm::asinhf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atan(x: f64) -> f64 {
-    libm::atan(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atan2(x: f64, y: f64) -> f64 {
-    libm::atan2(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atan2f(x: f32, y: f32) -> f32 {
-    libm::atan2f(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atanf(x: f32) -> f32 {
-    libm::atanf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atanh(x: f64) -> f64 {
-    libm::atanh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn atanhf(x: f32) -> f32 {
-    libm::atanhf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn cbrt(x: f64) -> f64 {
-    libm::cbrt(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn cbrtf(x: f32) -> f32 {
-    libm::cbrtf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ceil(x: f64) -> f64 {
-    libm::ceil(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ceilf(x: f32) -> f32 {
-    libm::ceilf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn copysign(x: f64, y: f64) -> f64 {
-    libm::copysign(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn copysignf(x: f32, y: f32) -> f32 {
-    libm::copysignf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn cos(x: f64) -> f64 {
-    libm::cos(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn cosf(x: f32) -> f32 {
-    libm::cosf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn cosh(x: f64) -> f64 {
-    libm::cosh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn coshf(x: f32) -> f32 {
-    libm::coshf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn erf(x: f64) -> f64 {
-    libm::erf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn erfc(x: f64) -> f64 {
-    libm::erfc(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn erfcf(x: f32) -> f32 {
-    libm::erfcf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn erff(x: f32) -> f32 {
-    libm::erff(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn exp(x: f64) -> f64 {
-    libm::exp(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn exp2(x: f64) -> f64 {
-    libm::exp2(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn exp2f(x: f32) -> f32 {
-    libm::exp2f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn exp10(x: f64) -> f64 {
-    libm::exp10(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn exp10f(x: f32) -> f32 {
-    libm::exp10f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn expf(x: f32) -> f32 {
-    libm::expf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn expm1(x: f64) -> f64 {
-    libm::expm1(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn expm1f(x: f32) -> f32 {
-    libm::expm1f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fabs(x: f64) -> f64 {
-    libm::fabs(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fabsf(x: f32) -> f32 {
-    libm::fabsf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fdim(x: f64, y: f64) -> f64 {
-    libm::fdim(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fdimf(x: f32, y: f32) -> f32 {
-    libm::fdimf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn floor(x: f64) -> f64 {
-    libm::floor(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn floorf(x: f32) -> f32 {
-    libm::floorf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fma(x: f64, y: f64, z: f64) -> f64 {
-    libm::fma(x, y, z)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmaf(x: f32, y: f32, z: f32) -> f32 {
-    libm::fmaf(x, y, z)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmax(x: f64, y: f64) -> f64 {
-    libm::fmax(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmaxf(x: f32, y: f32) -> f32 {
-    libm::fmaxf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmin(x: f64, y: f64) -> f64 {
-    libm::fmin(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fminf(x: f32, y: f32) -> f32 {
-    libm::fminf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmod(x: f64, y: f64) -> f64 {
-    libm::fmod(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn fmodf(x: f32, y: f32) -> f32 {
-    libm::fmodf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn frexp(x: f64, y: *mut i32) -> f64 {
-    let (a, b) = libm::frexp(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn frexpf(x: f32, y: *mut i32) -> f32 {
-    let (a, b) = libm::frexpf(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn hypot(x: f64, y: f64) -> f64 {
-    libm::hypot(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn hypotf(x: f32, y: f32) -> f32 {
-    libm::hypotf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ilogb(x: f64) -> i32 {
-    libm::ilogb(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ilogbf(x: f32) -> i32 {
-    libm::ilogbf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn j0(x: f64) -> f64 {
-    libm::j0(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn j0f(x: f32) -> f32 {
-    libm::j0f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn j1(x: f64) -> f64 {
-    libm::j1(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn j1f(x: f32) -> f32 {
-    libm::j1f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn jn(x: i32, y: f64) -> f64 {
-    libm::jn(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn jnf(x: i32, y: f32) -> f32 {
-    libm::jnf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ldexp(x: f64, y: i32) -> f64 {
-    libm::ldexp(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ldexpf(x: f32, y: i32) -> f32 {
-    libm::ldexpf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn lgamma_r(x: f64, y: *mut i32) -> f64 {
-    let (a, b) = libm::lgamma_r(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn lgammaf_r(x: f32, y: *mut i32) -> f32 {
-    let (a, b) = libm::lgammaf_r(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn log(x: f64) -> f64 {
-    libm::log(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log1p(x: f64) -> f64 {
-    libm::log1p(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log1pf(x: f32) -> f32 {
-    libm::log1pf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log2(x: f64) -> f64 {
-    libm::log2(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log2f(x: f32) -> f32 {
-    libm::log2f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log10(x: f64) -> f64 {
-    libm::log10(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn log10f(x: f32) -> f32 {
-    libm::log10f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn logf(x: f32) -> f32 {
-    libm::logf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn modf(x: f64, y: *mut f64) -> f64 {
-    let (a, b) = libm::modf(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn modff(x: f32, y: *mut f32) -> f32 {
-    let (a, b) = libm::modff(x);
-    *y = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn nextafter(x: f64, y: f64) -> f64 {
-    libm::nextafter(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn nextafterf(x: f32, y: f32) -> f32 {
-    libm::nextafterf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn pow(x: f64, y: f64) -> f64 {
-    libm::pow(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn powf(x: f32, y: f32) -> f32 {
-    libm::powf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn remainder(x: f64, y: f64) -> f64 {
-    libm::remainder(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn remainderf(x: f32, y: f32) -> f32 {
-    libm::remainderf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn remquo(x: f64, y: f64, z: *mut i32) -> f64 {
-    let (a, b) = libm::remquo(x, y);
-    *z = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn remquof(x: f32, y: f32, z: *mut i32) -> f32 {
-    let (a, b) = libm::remquof(x, y);
-    *z = b;
-    a
-}
-
-#[no_mangle]
-unsafe extern "C" fn round(x: f64) -> f64 {
-    libm::round(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn roundf(x: f32) -> f32 {
-    libm::roundf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn scalbn(x: f64, y: i32) -> f64 {
-    libm::scalbn(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn scalbnf(x: f32, y: i32) -> f32 {
-    libm::scalbnf(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sin(x: f64) -> f64 {
-    libm::sin(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sincos(x: f64, y: *mut f64, z: *mut f64) {
-    let (a, b) = libm::sincos(x);
-    *y = a;
-    *z = b;
-}
-
-#[no_mangle]
-unsafe extern "C" fn sincosf(x: f32, y: *mut f32, z: *mut f32) {
-    let (a, b) = libm::sincosf(x);
-    *y = a;
-    *z = b;
-}
-
-#[no_mangle]
-unsafe extern "C" fn sinf(x: f32) -> f32 {
-    libm::sinf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sinh(x: f64) -> f64 {
-    libm::sinh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sinhf(x: f32) -> f32 {
-    libm::sinhf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sqrt(x: f64) -> f64 {
-    libm::sqrt(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn sqrtf(x: f32) -> f32 {
-    libm::sqrtf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tan(x: f64) -> f64 {
-    libm::tan(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tanf(x: f32) -> f32 {
-    libm::tanf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tanh(x: f64) -> f64 {
-    libm::tanh(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tanhf(x: f32) -> f32 {
-    libm::tanhf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tgamma(x: f64) -> f64 {
-    libm::tgamma(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn tgammaf(x: f32) -> f32 {
-    libm::tgammaf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn trunc(x: f64) -> f64 {
-    libm::trunc(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn truncf(x: f32) -> f32 {
-    libm::truncf(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn y0(x: f64) -> f64 {
-    libm::y0(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn y0f(x: f32) -> f32 {
-    libm::y0f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn y1(x: f64) -> f64 {
-    libm::y1(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn y1f(x: f32) -> f32 {
-    libm::y1f(x)
-}
-
-#[no_mangle]
-unsafe extern "C" fn yn(x: i32, y: f64) -> f64 {
-    libm::yn(x, y)
-}
-
-#[no_mangle]
-unsafe extern "C" fn ynf(x: i32, y: f32) -> f32 {
-    libm::ynf(x, y)
 }
 
 // Environment variables
