+++ c-scape/src/mem/mod.rs
@@ -1,0 +1,3 @@
+mod bstring;
+
+pub use bstring::*;
